pyproject.toml
[project]
name = "arkhon-rheo"
version = "0.1.0"
description = "A sophisticated multi-agent orchestration framework built on LangGraph and Google GenAI."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Yao-Chu Wu", email = "yaochuwu.tw@gmail.com" },
]
maintainers = [
    { name = "Arkhon-Rheo Maintainers", email = "maintainers@arkhon-rheo.ai" },
]
keywords = ["multi-agent", "orchestration", "langgraph", "gemini", "ai-agent", "tdd"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "gemma>=3.3.0",
    "google-genai>=1.63.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.8",
    "langsmith>=0.7.3",
    "ollama>=0.6.1",
    "prometheus-client>=0.24.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
    "click>=8.1.7",
    "dotenv>=0.9.9",
]

[project.scripts]
arkhon-rheo = "arkhon_rheo.cli.main:main"

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "radon>=6.0.1",
    "ruff>=0.15.1",
    "ty>=0.0.17",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
norecursedirs = ["repos"]

[tool.ruff]
line-length = 120
target-version = "py312"
fix = true

[tool.ruff.lint]
# Enable all standard rules plus specific extras
select = [
    "E", "F", "W",   # pycodestyle, pyflakes
    "I",             # isort
    "N",             # pep8-naming
    "UP",            # pyupgrade
    "B",             # flake8-bugbear
    "S",             # flake8-bandit
    "A",             # flake8-builtins
    "C4",            # flake8-comprehensions
    "T20",           # flake8-print
    "RET",           # flake8-return
    "SIM",           # flake8-simplify
    "ARG",           # flake8-unused-arguments
    "PTH",           # flake8-use-pathlib
    "ERA",           # eradicate
    "PL",            # pylint
    "RUF",           # ruff-specific
]
ignore = [
    "PLR0913", # Allow many arguments
    "PLR2004", # Allow magic values in tests/demonstrations
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["arkhon_rheo"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "ARG002", "A002"]
"src/arkhon_rheo/tools/target_project.py" = ["S602"]
"src/arkhon_rheo/tools/builtin/calculator.py" = ["S307"]

[tool.ty.environment]
python-version = "3.12"
python-platform = "linux"
root = ["./src"]

[tool.ty.analysis]
respect-type-ignore-comments = true
allowed-unresolved-imports = [
    "google.genai.*",
    "langchain.*",
    "langgraph.*"
]

[tool.ty.rules]
# Suppress specific false positives for LangGraph state management
invalid-argument-type = "warn"
possibly-unresolved-reference = "warn"
unresolved-import = "error"
all = "warn"
